# Conflicts:
#	build.gradle.kts
#	src/commonMain/kotlin/exchange.dydx.abacus/state/internalstate/InternalState.kt
#	v4_abacus.podspec

+++ src/commonMain/kotlin/indexer/codegen/IndexerFillResponseObject.kt
@@ -46,5 +46,5 @@ data class IndexerFillResponseObject(
     val createdAtHeight: kotlin.String? = null,
     val orderId: kotlin.String? = null,
     val clientMetadata: kotlin.String? = null,
-    val subaccountNumber: kotlin.Double? = null
+    val subaccountNumber: kotlin.Int? = null
 )
